Moves Murmur3_x86_32 and Platform to package o.a.s.util.sketch.unsafe
